[HSC-416] fix: customer 런타임 async executor 등록 경로 고정
